try service strategy

+++ app/address/addressDetail.js
@@ -4,7 +4,7 @@
 'use strict';
 
 angular.module("address.addressDetail",[])
-.controller("addressDetailCtrl",['$scope','BASIC_EVENTS','addressService',function($scope,BASIC_EVENTS,addressService){
+.controller("addressDetailCtrl",function($scope,BASIC_EVENTS,addressService,confirmationDialogService,SharedState){
     $scope.addressDetail = {
         addressId:0,
         calledName:'',
@@ -37,23 +37,37 @@ angular.module("address.addressDetail",[])
             });
     };
     $scope.addAddress = function(){
-        addressService.addAddress($scope.addressDetail)
-            .success(function(){
-                $scope.reLoad();
-                alert('新增成功！')
-            }).error(function(res){
-                $scope.reLoad();
-                alert('新增失败:'+ res.message);
-            });
+        addressService.addAddress($scope.addressDetail).then(function(){
+            $scope.reLoad();
+        });
+            //.success(function(){
+            //    $scope.reLoad();
+            //    alert('新增成功！')
+            //}).error(function(res){
+            //    $scope.reLoad();
+            //    alert('新增失败:'+ res.message);
+            //});
     };
     $scope.deleteAddress = function(){
-        addressService.deleteAddress($scope.addressDetail)
-            .success(function(){
-                $scope.reLoad();
-                alert('删除成功！')
-            }).error(function(res){
-                $scope.reLoad();
-                alert('删除失败:'+ res.message);
-            });
+        var custName = "lalalal";
+        var modalOptions = {
+            closeButtonText: '取消',
+            actionButtonText: '删除',
+            headerText: '删除地址？',
+            bodyText: '是否要删除地址：'+  $scope.addressDetail.addressDesc  +'？',
+            modalType:confirmationDialogService.modalTypes.CONFIRM_MODAL
+        };
+
+        confirmationDialogService.showModal(modalOptions).then(function (result) {
+            addressService.deleteAddress($scope.addressDetail)
+                .success(function(){
+                    $scope.reLoad();
+                    alert('删除成功！')
+                }).error(function(res){
+                    $scope.reLoad();
+                    alert('删除失败:'+ res.message);
+                });
+            SharedState.turnOff('addressDetailModal');
+        });
     };
-}]);
+});
